requirements.txt
Django>=1.8
djangorestframework
